requirements.txt
streamlit==1.16.0
matplotlib==3.4.3
numpy==1.21.2
scipy==1.7.1
Pillow==8.3.1
pandas==1.3.3
scikit-learn==0.24.2
